@@ -6,34 +6,13 @@
 
 This module provides the main EEGDash class which serves as the primary entry point for
 interacting with the EEGDash ecosystem. It offers methods to query, insert, and update
-metadata records stored in the EEGDash database via REST API, and includes utilities to
-load EEG data from S3 for matched records.
+metadata records stored in the EEGDash database via REST API.
 """
 
-import json
-import os
-from pathlib import Path
 from typing import Any, Mapping
 
-import numpy as np
-import pandas as pd
-
-from .bids_eeg_metadata import (
-    build_query_from_kwargs,
-    load_eeg_attrs_from_bids_file,
-)
-from .const import (
-    ALLOWED_QUERY_FIELDS,
-)
-from .const import config as data_config
-from .dataset.bids_dataset import EEGBIDSDataset
-from .dataset.dataset import EEGDashDataset
-from .http_api_client import HTTPAPIConnectionManager
-from .logging import logger
-
-# Default public API endpoint
-# Note: DNS override is applied in http_api_client.py via socket patch
-EEGDASH_API_URL = "https://data.eegdash.org"
+from .bids_metadata import merge_query
+from .http_api_client import get_client
 
 
 class EEGDash:
@@ -49,7 +28,7 @@ class EEGDash:
     def __init__(
         self,
         *,
-        is_staging: bool = False,
+        database: str = "eegdash",
         api_url: str | None = None,
         auth_token: str | None = None,
     ) -> None:
@@ -57,9 +36,10 @@ class EEGDash:
 
         Parameters
         ----------
-        is_staging : bool, default False
-            If ``True``, use the staging database (``eegdashstaging``); otherwise
-            use the production database (``eegdash``).
+        database : str, default "eegdash"
+            Name of the MongoDB database to connect to. Common values:
+            ``"eegdash"`` (production), ``"eegdash_staging"`` (staging),
+            ``"eegdash_v1"`` (legacy archive).
         api_url : str, optional
             Override the default API URL. If not provided, uses the default
             public endpoint or the ``EEGDASH_API_URL`` environment variable.
@@ -69,28 +49,12 @@ class EEGDash:
 
         Examples
         --------
-        >>> eegdash = EEGDash()
+        >>> eegdash = EEGDash()  # production
+        >>> eegdash = EEGDash(database="eegdash_staging")  # staging
         >>> records = eegdash.find({"dataset": "ds002718"})
 
         """
-        self.config = data_config
-        self.is_staging = is_staging
-        self._api_url = api_url
-        self._auth_token = auth_token
-        self._init_api_client()
-
-    def _init_api_client(self) -> None:
-        """Initialize HTTP API client connection."""
-        # Determine API URL: parameter > environment variable > default
-        url = self._api_url or os.getenv("EEGDASH_API_URL", EEGDASH_API_URL)
-
-        # Auth token is optional for public reads, only needed for admin writes
-        token = self._auth_token or os.getenv("EEGDASH_API_TOKEN")
-
-        # Use singleton to get HTTP API client, database, and collection
-        self.__client, self.__db, self.__collection = (
-            HTTPAPIConnectionManager.get_client(url, self.is_staging, token)
-        )
+        self._client = get_client(api_url, database, auth_token)
 
     def find(
         self, query: dict[str, Any] = None, /, **kwargs
@@ -99,6 +63,8 @@ class EEGDash:
 
         Examples
         --------
+        >>> from eegdash import EEGDash
+        >>> eegdash = EEGDash()
         >>> eegdash.find({"dataset": "ds002718", "subject": {"$in": ["012", "013"]}})  # pre-built query
         >>> eegdash.find(dataset="ds002718", subject="012")  # keyword filters
         >>> eegdash.find(dataset="ds002718", subject=["012", "013"])  # sequence -> $in
@@ -122,96 +88,37 @@ class EEGDash:
             DB records that match the query.
 
         """
-        # Extract pagination parameters before building query
         limit = kwargs.pop("limit", None)
         skip = kwargs.pop("skip", None)
+        final_query = merge_query(query, require_query=True, **kwargs)
+        find_kwargs = {
+            k: v for k, v in {"limit": limit, "skip": skip}.items() if v is not None
+        }
+        return list(self._client.find(final_query, **find_kwargs))
 
-        final_query: dict[str, Any] | None = None
-
-        # Accept explicit empty dict {} to mean "match all"
-        raw_query = query if isinstance(query, dict) else None
-        kwargs_query = build_query_from_kwargs(**kwargs) if kwargs else None
-
-        # Determine presence, treating {} as a valid raw query
-        has_raw = isinstance(raw_query, dict)
-        has_kwargs = kwargs_query is not None
-
-        if has_raw and has_kwargs:
-            # Detect conflicting constraints on the same field (e.g., task specified
-            # differently in both places) and raise a clear error instead of silently
-            # producing an empty result.
-            self._raise_if_conflicting_constraints(raw_query, kwargs_query)
-            # Merge with logical AND so both constraints apply
-            if raw_query:  # non-empty dict adds constraints
-                final_query = {"$and": [raw_query, kwargs_query]}
-            else:  # {} adds nothing; use kwargs_query only
-                final_query = kwargs_query
-        elif has_raw:
-            # May be {} meaning match-all, or a non-empty dict
-            final_query = raw_query
-        elif has_kwargs:
-            final_query = kwargs_query
-        else:
-            # Avoid accidental full scans
-            raise ValueError(
-                "find() requires a query dictionary or at least one keyword argument. "
-                "To find all documents, use find({})."
-            )
-
-        # Pass limit and skip to the collection's find method
-        find_kwargs = {}
-        if limit is not None:
-            find_kwargs["limit"] = limit
-        if skip is not None:
-            find_kwargs["skip"] = skip
-
-        results = self.__collection.find(final_query, **find_kwargs)
-
-        return list(results)
-
-    def exist(self, query: dict[str, Any]) -> bool:
-        """Return True if at least one record matches the query, else False.
-
-        This is a lightweight existence check that uses MongoDB's ``find_one``
-        instead of fetching all matching documents (which would be wasteful in
-        both time and memory for broad queries). Only a restricted set of
-        fields is accepted to avoid accidental full scans caused by malformed
-        or unsupported keys.
+    def exists(self, query: dict[str, Any] = None, /, **kwargs) -> bool:
+        """Check if at least one record matches the query.
 
         Parameters
         ----------
-        query : dict
-            Mapping of allowed field(s) to value(s). Allowed keys: ``data_name``
-            and ``dataset``. The query must not be empty.
+        query : dict, optional
+            Complete query dictionary. This is a positional-only argument.
+        **kwargs
+            User-friendly field filters (same as find()).
 
         Returns
         -------
         bool
             True if at least one matching record exists; False otherwise.
 
-        Raises
-        ------
-        TypeError
-            If ``query`` is not a dict.
-        ValueError
-            If ``query`` is empty or contains unsupported field names.
+        Examples
+        --------
+        >>> eeg = EEGDash()
+        >>> eeg.exists(dataset="ds002718")  # check by dataset
+        >>> eeg.exists({"data_name": "ds002718_sub-001_eeg.set"})  # check by data_name
 
         """
-        if not isinstance(query, dict):
-            raise TypeError("query must be a dict")
-        if not query:
-            raise ValueError("query cannot be empty")
-
-        accepted_query_fields = {"data_name", "dataset"}
-        unknown = set(query.keys()) - accepted_query_fields
-        if unknown:
-            raise ValueError(
-                f"Unsupported query field(s): {', '.join(sorted(unknown))}. "
-                f"Allowed: {sorted(accepted_query_fields)}"
-            )
-
-        doc = self.__collection.find_one(query, projection={"_id": 1})
-        return doc is not None
+        return self.find_one(query, **kwargs) is not None
 
     def count(self, query: dict[str, Any] = None, /, **kwargs) -> int:
         """Count documents matching the query.
@@ -235,360 +142,112 @@ class EEGDash:
         >>> count = eeg.count(dataset="ds002718")  # count by dataset
 
         """
-        # Extract limit/skip if present (not used for count but kept for consistency)
         kwargs.pop("limit", None)
         kwargs.pop("skip", None)
+        final_query = merge_query(query, require_query=False, **kwargs)
+        return self._client.count_documents(final_query)
 
-        final_query: dict[str, Any] | None = None
-
-        # Accept explicit empty dict {} to mean "match all"
-        raw_query = query if isinstance(query, dict) else None
-        kwargs_query = build_query_from_kwargs(**kwargs) if kwargs else None
-
-        # Determine presence, treating {} as a valid raw query
-        has_raw = isinstance(raw_query, dict)
-        has_kwargs = kwargs_query is not None
-
-        if has_raw and has_kwargs:
-            self._raise_if_conflicting_constraints(raw_query, kwargs_query)
-            if raw_query:
-                final_query = {"$and": [raw_query, kwargs_query]}
-            else:
-                final_query = kwargs_query
-        elif has_raw:
-            final_query = raw_query
-        elif has_kwargs:
-            final_query = kwargs_query
-        else:
-            # For count, empty query is acceptable (count all)
-            final_query = {}
-
-        return self.__collection.count_documents(final_query)
-
-    def _validate_input(self, record: dict[str, Any]) -> dict[str, Any]:
-        """Validate the input record against the expected schema.
+    def find_one(
+        self, query: dict[str, Any] = None, /, **kwargs
+    ) -> Mapping[str, Any] | None:
+        """Find a single record matching the query.
 
         Parameters
         ----------
-        record : dict
-            A dictionary representing the EEG data record to be validated.
-
-        Returns
-        -------
-        dict
-            The record itself on success.
-
-        Raises
-        ------
-        ValueError
-            If the record is missing required keys or has values of the wrong type.
-
-        """
-        input_types = {
-            "data_name": str,
-            "dataset": str,
-            "bidspath": str,
-            "subject": str,
-            "task": str,
-            "session": str,
-            "run": str,
-            "sampling_frequency": float,
-            "modality": str,
-            "nchans": int,
-            "ntimes": int,
-            "channel_types": list,
-            "channel_names": list,
-        }
-        if "data_name" not in record:
-            raise ValueError("Missing key: data_name")
-        # check if args are in the keys and has correct type
-        for key, value in record.items():
-            if key not in input_types:
-                raise ValueError(f"Invalid input: {key}")
-            if not isinstance(value, input_types[key]):
-                raise ValueError(f"Invalid input: {key}")
-
-        return record
-
-    def _build_query_from_kwargs(self, **kwargs) -> dict[str, Any]:
-        """Build a validated MongoDB query from keyword arguments.
-
-        This delegates to the module-level builder used across the package.
-
-        Parameters
-        ----------
+        query : dict, optional
+            Complete query dictionary. This is a positional-only argument.
         **kwargs
-            Keyword arguments to convert into a MongoDB query.
+            User-friendly field filters (same as find()).
 
         Returns
         -------
-        dict
-            A MongoDB query dictionary.
+        dict or None
+            The first matching record, or None if no match.
+
+        Examples
+        --------
+        >>> eeg = EEGDash()
+        >>> record = eeg.find_one(data_name="ds002718_sub-001_eeg.set")
 
         """
-        return build_query_from_kwargs(**kwargs)
+        final_query = merge_query(query, require_query=True, **kwargs)
+        return self._client.find_one(final_query)
 
-    def _extract_simple_constraint(
-        self, query: dict[str, Any], key: str
-    ) -> tuple[str, Any] | None:
-        """Extract a simple constraint for a given key from a query dict.
-
-        Supports top-level equality (e.g., ``{'subject': '01'}``) and ``$in``
-        (e.g., ``{'subject': {'$in': ['01', '02']}}``) constraints.
+    def insert(self, records: dict[str, Any] | list[dict[str, Any]]) -> int:
+        """Insert one or more records (requires auth_token).
 
         Parameters
         ----------
-        query : dict
-            The MongoDB query dictionary.
-        key : str
-            The key for which to extract the constraint.
+        records : dict or list of dict
+            A single record or list of records to insert.
 
         Returns
         -------
-        tuple or None
-            A tuple of (kind, value) where kind is "eq" or "in", or None if the
-            constraint is not present or unsupported.
+        int
+            Number of records inserted.
+
+        Examples
+        --------
+        >>> eeg = EEGDash(auth_token="...")
+        >>> eeg.insert({"dataset": "ds001", "subject": "01", ...})  # single
+        >>> eeg.insert([record1, record2, record3])  # batch
 
         """
-        if not isinstance(query, dict) or key not in query:
-            return None
-        val = query[key]
-        if isinstance(val, dict):
-            if "$in" in val and isinstance(val["$in"], (list, tuple)):
-                return ("in", list(val["$in"]))
-            return None  # unsupported operator shape for conflict checking
-        else:
-            return "eq", val
+        if isinstance(records, dict):
+            self._client.insert_one(records)
+            return 1
+        return self._client.insert_many(records)
 
-    def _raise_if_conflicting_constraints(
-        self, raw_query: dict[str, Any], kwargs_query: dict[str, Any]
-    ) -> None:
-        """Raise ValueError if query sources have incompatible constraints.
-
-        Checks for mutually exclusive constraints on the same field to avoid
-        silent empty results.
-
-        Parameters
-        ----------
-        raw_query : dict
-            The raw MongoDB query dictionary.
-        kwargs_query : dict
-            The query dictionary built from keyword arguments.
-
-        Raises
-        ------
-        ValueError
-            If conflicting constraints are found.
-
-        """
-        if not raw_query or not kwargs_query:
-            return
-
-        # Only consider fields we generally allow; skip meta operators like $and
-        raw_keys = set(raw_query.keys()) & ALLOWED_QUERY_FIELDS
-        kw_keys = set(kwargs_query.keys()) & ALLOWED_QUERY_FIELDS
-        dup_keys = raw_keys & kw_keys
-        for key in dup_keys:
-            rc = self._extract_simple_constraint(raw_query, key)
-            kc = self._extract_simple_constraint(kwargs_query, key)
-            if rc is None or kc is None:
-                # If either side is non-simple, skip conflict detection for this key
-                continue
-
-            r_kind, r_val = rc
-            k_kind, k_val = kc
-
-            # Normalize to sets when appropriate for simpler checks
-            if r_kind == "eq" and k_kind == "eq":
-                if r_val != k_val:
-                    raise ValueError(
-                        f"Conflicting constraints for '{key}': query={r_val!r} vs kwargs={k_val!r}"
-                    )
-            elif r_kind == "in" and k_kind == "eq":
-                if k_val not in r_val:
-                    raise ValueError(
-                        f"Conflicting constraints for '{key}': query in {r_val!r} vs kwargs={k_val!r}"
-                    )
-            elif r_kind == "eq" and k_kind == "in":
-                if r_val not in k_val:
-                    raise ValueError(
-                        f"Conflicting constraints for '{key}': query={r_val!r} vs kwargs in {k_val!r}"
-                    )
-            elif r_kind == "in" and k_kind == "in":
-                if len(set(r_val).intersection(k_val)) == 0:
-                    raise ValueError(
-                        f"Conflicting constraints for '{key}': disjoint sets {r_val!r} and {k_val!r}"
-                    )
-
-    def add_bids_dataset(
+    def update_field(
         self,
-        dataset: str,
-        data_dir: str,
-        overwrite: bool = True,
-        output_path: str | Path | None = None,
-    ) -> dict[str, Any]:
-        """Collect metadata for a local BIDS dataset as JSON-ready records.
+        query: dict[str, Any] = None,
+        /,
+        *,
+        update: dict[str, Any],
+        **kwargs,
+    ) -> tuple[int, int]:
+        """Update fields on records matching the query (requires auth_token).
 
-        Instead of inserting records directly into MongoDB, this method scans
-        ``data_dir`` and returns a JSON-serializable manifest describing every
-        EEG recording that was discovered. The manifest can be written to disk
-        or forwarded to the EEGDash ingestion API for persistence.
-
-        Parameters
-        ----------
-        dataset : str
-            Dataset identifier (e.g., ``"ds002718"``).
-        data_dir : str
-            Path to the local BIDS dataset directory.
-        overwrite : bool, default True
-            If ``False``, skip records that already exist in the database based
-            on ``data_name`` lookups.
-        output_path : str | Path | None, optional
-            If provided, the manifest is written to the given JSON file.
-
-        Returns
-        -------
-        dict
-            A manifest with keys ``dataset``, ``source``, ``records`` and, when
-            applicable, ``skipped`` or ``errors``.
-
-        """
-        source_dir = Path(data_dir).expanduser()
-        try:
-            bids_dataset = EEGBIDSDataset(
-                data_dir=str(source_dir),
-                dataset=dataset,
-            )
-        except Exception as exc:
-            logger.error("Error creating BIDS dataset %s: %s", dataset, exc)
-            raise exc
-
-        records: list[dict[str, Any]] = []
-        skipped: list[str] = []
-        errors: list[dict[str, str]] = []
-
-        for bids_file in bids_dataset.get_files():
-            data_id = f"{dataset}_{Path(bids_file).name}"
-            if not overwrite:
-                try:
-                    if self.exist({"data_name": data_id}):
-                        skipped.append(data_id)
-                        continue
-                except Exception as exc:
-                    logger.warning(
-                        "Could not verify existing record %s due to: %s",
-                        data_id,
-                        exc,
-                    )
-
-            try:
-                eeg_attrs = load_eeg_attrs_from_bids_file(bids_dataset, bids_file)
-                records.append(eeg_attrs)
-            except Exception as exc:  # log and continue collecting
-                logger.error("Error extracting metadata for %s", bids_file)
-                logger.error(str(exc))
-                errors.append({"file": str(bids_file), "error": str(exc)})
-
-        manifest: dict[str, Any] = {
-            "dataset": dataset,
-            "source": str(source_dir.resolve()),
-            "record_count": len(records),
-            "records": records,
-        }
-        if skipped:
-            manifest["skipped"] = skipped
-        if errors:
-            manifest["errors"] = errors
-
-        if output_path is not None:
-            output_path = Path(output_path)
-            output_path.parent.mkdir(parents=True, exist_ok=True)
-            with output_path.open("w", encoding="utf-8") as fh:
-                json.dump(
-                    manifest,
-                    fh,
-                    indent=2,
-                    sort_keys=True,
-                    default=_json_default,
-                )
-            logger.info(
-                "Wrote EEGDash ingestion manifest for %s to %s",
-                dataset,
-                output_path,
-            )
-
-        logger.info(
-            "Prepared %s records for dataset %s (skipped=%s, errors=%s)",
-            len(records),
-            dataset,
-            len(skipped),
-            len(errors),
-        )
-
-        return manifest
-
-    def exists(self, query: dict[str, Any]) -> bool:
-        """Check if at least one record matches the query.
-
-        This is an alias for :meth:`exist`.
+        Use this to add or modify fields across matching records,
+        e.g., after re-extracting entities with an improved algorithm.
 
         Parameters
         ----------
-        query : dict
-            MongoDB query to check for existence.
+        query : dict, optional
+            Filter query to match records. This is a positional-only argument.
+        update : dict
+            Fields to update. Keys are field names, values are new values.
+        **kwargs
+            User-friendly field filters (same as find()).
 
         Returns
         -------
-        bool
-            True if a matching record exists, False otherwise.
+        tuple of (matched_count, modified_count)
+            Number of records matched and actually modified.
+
+        Examples
+        --------
+        >>> eeg = EEGDash(auth_token="...")
+        >>> # Update entities for all records in a dataset
+        >>> eeg.update_field({"dataset": "ds002718"}, update={"entities": {"subject": "01"}})
+        >>> # Using kwargs for filter
+        >>> eeg.update_field(dataset="ds002718", update={"entities": new_entities})
+        >>> # Combine query + kwargs
+        >>> eeg.update_field({"dataset": "ds002718"}, subject="01", update={"entities": new_entities})
 
         """
-        return self.exist(query)
-
-    @property
-    def collection(self):
-        """The underlying collection interface.
-
-        Returns
-        -------
-        HTTPAPICollection
-            The collection object used for database interactions via REST API.
-
-        """
-        return self.__collection
-
-    @classmethod
-    def close_all_connections(cls) -> None:
-        """Close all HTTP API client connections managed by the singleton manager."""
-        HTTPAPIConnectionManager.close_all()
+        final_query = merge_query(query, require_query=True, **kwargs)
+        return self._client.update_many(final_query, update)
 
 
-def _json_default(value: Any) -> Any:
-    """Fallback serializer for complex objects when exporting ingestion JSON."""
-    try:
-        if isinstance(value, (np.generic,)):
-            return value.item()
-        if isinstance(value, np.ndarray):
-            return value.tolist()
-    except Exception:
-        pass
+def __getattr__(name: str):
+    # Backward-compat: allow ``from eegdash.api import EEGDashDataset`` without
+    # importing braindecode unless needed.
+    if name == "EEGDashDataset":
+        from .dataset.dataset import EEGDashDataset
 
-    try:
-        if value is pd.NA:
-            return None
-        if isinstance(value, (pd.Timestamp, pd.Timedelta)):
-            return value.isoformat()
-        if isinstance(value, pd.Series):
-            return value.to_dict()
-    except Exception:
-        pass
-
-    if isinstance(value, Path):
-        return value.as_posix()
-    if isinstance(value, set):
-        return sorted(value)
-
-    raise TypeError(f"Object of type {type(value).__name__} is not JSON serializable")
+        return EEGDashDataset
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
 
 
-__all__ = ["EEGDash", "EEGDashDataset"]
+__all__ = ["EEGDash"]
