@@ -54,5 +54,5 @@ def connectivity_imaginary_coherence(f, c, /, bands=utils.DEFAULT_FREQ_BANDS):
 @bivariate_feature
 def connectivity_lagged_coherence(f, c, /, bands=utils.DEFAULT_FREQ_BANDS):
     # https://neuroimage.usc.edu/brainstorm/Tutorials/Connectivity
-    coher = c.imag / np.sqrt(1 - c.real)
+    coher = c.imag / np.sqrt(1 - c.real**2)
     return utils.reduce_freq_bands(f, coher, bands, np.mean)
