@@ -9,11 +9,15 @@
    :depth: 2
 """
 
+#
 ######################################################################
+#
 # .. image:: https://colab.research.google.com/assets/colab-badge.svg
 #    :target: https://colab.research.google.com/github/eeg2025/startkit/blob/main/challenge_1.ipynb
 #    :alt: Open In Colab
+#
 ######################################################################
+#
 # Preliminary notes
 # -----------------
 # Before we begin, I just want to make a deal with you, ok?
@@ -25,7 +29,9 @@
 #
 # The entire decoding community will only go further when we stop
 # solving the same problems over and over again, and it starts working together.
+#
 ######################################################################
+#
 # How can we use the knowledge from one EEG Decoding task into another?
 # ---------------------------------------------------------------------
 # Transfer learning is a widespread technique used in deep learning. It
@@ -55,13 +61,17 @@
 # This means your submitted model might be trained on a subset of tasks
 # and fine-tuned on data from another condition, evaluating its capacity to
 # generalize with task-specific fine-tuning.
+#
 ######################################################################
+#
 # __________
 #
 # Note: For simplicity purposes, we will only show how to do the decoding
 # directly in our target task, and it is up to the teams to think about
 # how to use the passive task to perform the pre-training.
+#
 #######################################################################
+#
 # Install dependencies
 # --------------------
 # For the challenge, we will need two significant dependencies:
@@ -72,10 +82,13 @@
 # Install dependencies on colab or your local machine, as eegdash
 # have braindecode as a dependency.
 # you can just run ``pip install eegdash``.
+#
 ######################################################################
+#
 # Imports and setup
 # -----------------
 from pathlib import Path
+
 import torch
 from braindecode.datasets import BaseConcatDataset
 from braindecode.preprocessing import (
@@ -92,9 +105,10 @@ from torch.nn import Module
 from torch.optim.lr_scheduler import LRScheduler
 from tqdm import tqdm
 import copy
-from joblib import Parallel, delayed
 
+#
 ######################################################################
+#
 # Check GPU availability
 # ----------------------
 #
@@ -114,7 +128,9 @@ else:
         "selecting 'T4 GPU'\nunder 'Hardware accelerator'."
     )
 print(msg)
+#
 ######################################################################
+#
 # What are we decoding?
 # ---------------------
 #
@@ -122,6 +138,7 @@ print(msg)
 # is to understand some basic concepts.
 #
 ######################################################################
+#
 # The brain decodes the problem
 # -----------------------------
 #
@@ -158,7 +175,9 @@ print(msg)
 # In this tutorial, we will use the ``EEGNeX`` model from ``braindecode`` as an example.
 # You can use any model you want, as long as it follows the input/output
 # definitions above.
+#
 ######################################################################
+#
 # Understand the task: Contrast Change Detection (CCD)
 # --------------------------------------------------------
 # If you are interested to get more neuroscience insight, we recommend these two references, `HBN-EEG <https://www.biorxiv.org/content/10.1101/2024.10.03.615261v2.full.pdf>`__ and `Langer, N et al. (2017) <https://www.nature.com/articles/sdata201740#Sec2>`__.
@@ -185,10 +204,13 @@ print(msg)
 # Your task (**label**) is to predict the response time for the subject during this windows.
 #
 #######################################################################
+#
 # In the figure below, we have the timeline representation of the cognitive task:
 #
 # .. image:: https://eeg2025.github.io/assets/img/image-2.jpg
+#
 ######################################################################
+#
 # Stimulus demonstration
 # ----------------------
 # .. raw:: html
@@ -201,6 +223,7 @@ print(msg)
 #    </div>
 #
 ######################################################################
+#
 # PyTorch Dataset for the competition
 # -----------------------------------
 # Now, we have a Pytorch Dataset object that contains the set of recordings for the task
@@ -213,9 +236,10 @@ from eegdash.hbn.windows import (
     keep_only_recordings_with,
     add_extras_columns,
 )
+from eegdash.paths import get_default_cache_dir
 
 # Match tests' cache layout under ~/eegdash_cache/eeg_challenge_cache
-DATA_DIR = (Path.home() / "eegdash_cache" / "eeg_challenge_cache").resolve()
+DATA_DIR = Path(get_default_cache_dir()).resolve()
 DATA_DIR.mkdir(parents=True, exist_ok=True)
 dataset_ccd = EEGChallengeDataset(
     task="contrastChangeDetection", release="R5", cache_dir=DATA_DIR, mini=True
@@ -235,8 +259,10 @@ print(
 # We can also access the Raw object for visualization purposes, we will see just one object.
 raw = dataset_ccd.datasets[0].raw  # get the Raw object of the first recording
 # And to download all the data all data directly, you can do:
-raws = Parallel(n_jobs=-1)(delayed(lambda d: d.raw)(d) for d in dataset_ccd.datasets)
+dataset_ccd.download_all(n_jobs=-1)
+#
 ######################################################################
+#
 # Alternatives for Downloading the data
 # -------------------------------------
 #
@@ -252,7 +278,9 @@ raws = Parallel(n_jobs=-1)(delayed(lambda d: d.raw)(d) for d in dataset_ccd.data
 #
 #    aws s3 sync s3://nmdatasets/NeurIPS25/R1_L100_bdf data/R1_L100_bdf --no-sign-request
 
+#
 ######################################################################
+#
 # Create windows of interest
 # -----------------------------
 # So we epoch after the stimulus moment with a beginning shift of 500 ms.
@@ -300,7 +328,9 @@ single_windows = add_extras_columns(
         "response_type",
     ),
 )
+#
 ######################################################################
+#
 # Inspect the label distribution
 # -------------------------------
 import numpy as np
@@ -318,7 +348,9 @@ ax.set_ylabel("Count")
 plt.tight_layout()
 plt.show()
 
+#
 ######################################################################
+#
 # Split the data
 # ---------------
 # Extract meta information
@@ -360,7 +392,9 @@ print("Number of examples in each split in the minirelease")
 print(f"Train:\t{len(train_set)}")
 print(f"Valid:\t{len(valid_set)}")
 print(f"Test:\t{len(test_set)}")
+#
 ######################################################################
+#
 # Create dataloaders
 # -------------------
 batch_size = 128
@@ -375,7 +409,9 @@ valid_loader = DataLoader(
 test_loader = DataLoader(
     test_set, batch_size=batch_size, shuffle=False, num_workers=num_workers
 )
+#
 ######################################################################
+#
 # Build the model
 # -----------------
 # For neural network models, **to start**, we suggest using `braindecode models <https://braindecode.org/1.2/models/models_table.html>`__ zoo.
@@ -393,7 +429,9 @@ model = EEGNeX(
 print(model)
 model.to(device)
 
+#
 ######################################################################
+#
 # Define training and validation functions
 # -------------------------------------------
 # The rest is our classic PyTorch/torch lighting/skorch training pipeline,
@@ -486,7 +524,9 @@ def valid_model(
     return avg_loss, rmse
 
 
+#
 ######################################################################
+#
 # Train the model
 # ------------------
 lr = 1e-3
@@ -529,7 +569,9 @@ for epoch in range(1, n_epochs + 1):
             break
 if best_state is not None:
     model.load_state_dict(best_state)
+#
 ######################################################################
+#
 # Save the model
 # -----------------
 torch.save(model.state_dict(), "weights_challenge_1.pt")
