@@ -10,10 +10,10 @@ present on disk.
 """
 
 from pathlib import Path
-import platformdirs
 
 from eegdash.const import RELEASE_TO_OPENNEURO_DATASET_MAP
-from eegdash.dataset.dataset import EEGChallengeDataset
+from eegdash.paths import get_default_cache_dir
+from eegdash import EEGChallengeDataset
 
 
 # We'll use Release R2 as an example (HBN subset).
@@ -23,10 +23,12 @@ release = "R2"
 dataset_id = RELEASE_TO_OPENNEURO_DATASET_MAP[release]
 task = "RestingState"
 # Choose a cache directory. This should be on a fast local filesystem.
-cache_dir = Path(platformdirs.user_cache_dir("EEGDash"))
+cache_dir = Path(get_default_cache_dir()).resolve()
 cache_dir.mkdir(parents=True, exist_ok=True)
 
+#
 #######################################################################
+#
 # Step 1: Populate the local cache (Online)
 # -----------------------------------------
 # This block downloads the dataset from S3 to your local cache directory.
@@ -44,12 +46,12 @@ ds_online = EEGChallengeDataset(
 )
 
 # Optional prefetch of all recordings (downloads everything to cache).
-from joblib import Parallel, delayed
-
-_ = Parallel(n_jobs=-1)(delayed(lambda d: d.raw)(d) for d in ds_online.datasets)
+ds_online.download_all(n_jobs=-1)
 
 
+#
 #######################################################################
+#
 # Step 2: Basic Offline Usage
 # ---------------------------
 # Once the data is cached locally, you can interact with it without needing an
@@ -76,7 +78,9 @@ if ds_offline.datasets:
     print("First record bidspath:", ds_offline.datasets[0].record["bidspath"])
 
 
+#
 #######################################################################
+#
 # Step 3: Filtering Entities Offline
 # ----------------------------------
 # Even without a database connection, you can still filter your dataset by
@@ -102,7 +106,9 @@ if ds_offline_sub.datasets:
         print(f"  {idx:03d}: {summary}")
 
 
+#
 #######################################################################
+#
 # Step 4: Comparing Online vs. Offline Data
 # -----------------------------------------
 # As a sanity check, you can verify that the data loaded from your local cache
@@ -120,7 +126,9 @@ print("online shape:", raw_online.get_data().shape)
 print("offline shape:", raw_offline.get_data().shape)
 print("shapes equal:", raw_online.get_data().shape == raw_offline.get_data().shape)
 
+#
 #######################################################################
+#
 # Step 4.1: Comparing Descriptions, Online vs. Offline Data
 # ---------------------------------------------------------
 #
@@ -134,7 +142,9 @@ print("Online description shape:", description_online.shape)
 print("Offline description shape:", description_offline.shape)
 print("Descriptions equal:", description_online.equals(description_offline))
 
+#
 #######################################################################
+#
 # Notes and troubleshooting
 # -------------------------
 # - Working offline selects recordings by parsing BIDS filenames and directory
